# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Expense & Budget Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [dashboard]
      summary: Get spending grouped by category
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Month in YYYY-MM format (default: current month)"
      responses:
        "200":
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpendingByCategory"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [dashboard]
      summary: Get monthly spending trend (last 6 months)
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrend"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dashboard/recent-expenses:
    get:
      operationId: getRecentExpenses
      tags: [dashboard]
      summary: Get recent expenses for activity feed
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Recent expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseWithDetails"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List user expenses with pagination and filters
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: budgetId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated list of expenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Expense deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        categoryId:
          type: ["integer", "null"]
        budgetId:
          type: ["integer", "null"]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date-time
        paymentMethod:
          type: string
          enum: [cash, credit_card, debit_card, bank_transfer, other]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - description
        - date
        - paymentMethod
        - createdAt
    ExpenseWithDetails:
      allOf:
        - $ref: "#/components/schemas/Expense"
        - type: object
          properties:
            categoryName:
              type: ["string", "null"]
            categoryColor:
              type: ["string", "null"]
            budgetName:
              type: ["string", "null"]
    ExpenseListResponse:
      type: object
      properties:
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseWithDetails"
        total:
          type: integer
        skip:
          type: integer
        limit:
          type: integer
      required:
        - expenses
        - total
        - skip
        - limit
    CreateExpenseBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        budgetId:
          type: ["integer", "null"]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date-time
        paymentMethod:
          type: string
          enum: [cash, credit_card, debit_card, bank_transfer, other]
      required:
        - amount
        - description
        - date
        - paymentMethod
    UpdateExpenseBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        budgetId:
          type: ["integer", "null"]
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date-time
        paymentMethod:
          type: string
          enum: [cash, credit_card, debit_card, bank_transfer, other]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [budget_alert, expense_added, report_ready, system]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - message
        - type
        - isRead
        - createdAt
    DashboardSummary:
      type: object
      properties:
        totalExpensesThisMonth:
          type: number
        totalExpensesLastMonth:
          type: number
        totalBudget:
          type: number
        budgetUsedPercent:
          type: number
        expenseCount:
          type: integer
        budgetCount:
          type: integer
        categoryCount:
          type: integer
        unreadNotifications:
          type: integer
      required:
        - totalExpensesThisMonth
        - totalExpensesLastMonth
        - totalBudget
        - budgetUsedPercent
        - expenseCount
        - budgetCount
        - categoryCount
        - unreadNotifications
    SpendingByCategory:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        categoryColor:
          type: string
        total:
          type: number
        count:
          type: integer
        percentage:
          type: number
      required:
        - categoryName
        - categoryColor
        - total
        - count
        - percentage
    MonthlyTrend:
      type: object
      properties:
        month:
          type: string
        total:
          type: number
        count:
          type: integer
      required:
        - month
        - total
        - count