# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense & Budget Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalExpensesThisMonth: number;
  totalExpensesLastMonth: number;
  totalBudget: number;
  budgetUsedPercent: number;
  expenseCount: number;
  budgetCount: number;
  categoryCount: number;
  unreadNotifications: number;
}