# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense & Budget Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpensePaymentMethod } from "./expensePaymentMethod";

export interface Expense {
  id: number;
  userId: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  budgetId?: number | null;
  amount: number;
  description: string;
  date: Date;
  paymentMethod: ExpensePaymentMethod;
  createdAt: Date;
}