# lib/api-zod/src/generated/types/expenseListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Expense & Budget Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseWithDetails } from "./expenseWithDetails";

export interface ExpenseListResponse {
  expenses: ExpenseWithDetails[];
  total: number;
  skip: number;
  limit: number;
}